+++ projects/frontend/es-app/src/guis/menus/GuiMenuUserScripts.cpp
@@ -7,24 +7,27 @@
 #include <usernotifications/NotificationManager.h>
 
 GuiMenuUserScripts::GuiMenuUserScripts(WindowManager& window)
-  : GuiMenuBase(window, "USER SCRIPTS", this)
+  : Menu(window, "USER SCRIPTS")
   , mSender(*this)
   , mWaiter(nullptr)
   , mScriptIndexToRun(0)
 {
-  int index = 0;
-  Path::PathList scripts = NotificationManager::Instance().GetManualScriptList();
-  std::sort(scripts.begin(), scripts.end());
-  for(const Path& path : scripts)
-    AddSubMenu(ExtractScriptName(path), index++);
+  NotificationManager::ScriptDescriptorList scripts = NotificationManager::Instance().GetManualScriptList();
+  std::sort(scripts.begin(), scripts.end(), [](const NotificationManager::ScriptDescriptor& a, NotificationManager::ScriptDescriptor& b){ return a.mPath.ToString() < b.mPath.ToString(); });
+  for(const NotificationManager::ScriptDescriptor& script : scripts)
+    AddAction(ExtractScriptName(script.mPath, script.mAttribute), _("RUN"), script.mIndex, true, this);
 }
 
-String GuiMenuUserScripts::ExtractScriptName(const Path& script)
+String GuiMenuUserScripts::ExtractScriptName(const Path& script, ScriptAttributes attributes)
 {
-  return GameFilesUtils::RemoveParenthesis(script.FilenameWithoutExtension());
+  String name = GameFilesUtils::RemoveParenthesis(script.FilenameWithoutExtension());
+  if ((attributes & (ScriptAttributes::Synced | ScriptAttributes::ReportProgress)) != 0) name.Append(' ');
+  if (hasFlag(attributes, ScriptAttributes::Synced)) name.AppendUTF8(0xF1AF);
+  if (hasFlag(attributes, ScriptAttributes::ReportProgress)) name.AppendUTF8(0xF1AE);
+  return name;
 }
 
-void GuiMenuUserScripts::SubMenuSelected(int id)
+void GuiMenuUserScripts::MenuActionTriggered(int id)
 {
   mScriptIndexToRun = id;
   Thread::Start("script");
@@ -71,10 +74,10 @@ void GuiMenuUserScripts::ReceiveSyncMessage()
       case ScriptEvent::Start:
       {
         if ((data.mAttributes & ScriptAttributes::Synced) == 0)
-          mWindow.displayMessage((_F(_("Script {0} started successfully!")) / ExtractScriptName(data.mPath)).ToString());
+          mWindow.displayMessage((_F(_("Script {0} started successfully!")) / ExtractScriptName(data.mPath, data.mAttributes)).ToString());
         else
         {
-          mWaiter = new GuiASyncWaiter(mWindow, (_F(_("Running {0}...")) / ExtractScriptName(data.mPath)).ToString());
+          mWaiter = new GuiASyncWaiter(mWindow, (_F(_("Running {0}...")) / ExtractScriptName(data.mPath, data.mAttributes)).ToString());
           mWindow.pushGui(mWaiter);
         }
         break;
@@ -107,14 +110,14 @@ void GuiMenuUserScripts::ReceiveSyncMessage()
           String title = _("Script execution complete");
           if (data.mError)
           {
-            String text = (_F(_("Script {0} has failed!")) / ExtractScriptName(data.mPath)).ToString();
+            String text = (_F(_("Script {0} has failed!")) / ExtractScriptName(data.mPath, data.mAttributes)).ToString();
             if (!data.mStdErr.empty())
-              text.Append('\n').Append(_("Width the following Error output:")).Append('\n').Append('\n').Append(data.mStdErr);
+              text.Append('\n').Append(_("With the following Error output:")).Append('\n').Append('\n').Append(data.mStdErr);
             mWindow.displayScrollMessage(title, text);
           }
-          String text = (_F(_("Script {0} executed successfully!")) / ExtractScriptName(data.mPath)).ToString();
+          String text = (_F(_("Script {0} executed successfully!")) / ExtractScriptName(data.mPath, data.mAttributes)).ToString();
           if (!hasFlag(data.mAttributes, ScriptAttributes::ReportProgress) && !data.mStdOut.empty())
-            text.Append('\n').Append(_("Width the following output:")).Append('\n').Append('\n').Append(data.mStdOut);
+            text.Append('\n').Append(_("With the following output:")).Append('\n').Append('\n').Append(data.mStdOut);
           mWindow.displayScrollMessage(title, text);
         }
         break;