+++ projects/frontend/es-app/src/usernotifications/NotificationManager.h
@@ -283,6 +283,17 @@ class NotificationManager : public StaticLifeCycleControler<NotificationManager>
     void Run() override;
 
   public:
+    /*!
+     * @brief Script descriptor
+     */
+    struct ScriptDescriptor
+    {
+      Path             mPath;      //!< Script path
+      int              mIndex;     //! internal index required for manual script execution
+      ScriptAttributes mAttribute; //!< Script attributes
+    };
+    //! Script descriptor list
+    typedef std::vector<ScriptDescriptor> ScriptDescriptorList;
 
     /*!
      * @brief Default constructor